--- Task6/Task6/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Task6
-//
-//  Created by Admin on 24.11.2020.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-	override func viewDidLoad() {
-		super.viewDidLoad()
-		// Do any additional setup after loading the view.
-	}
-
-
-}
-
